# perf(remix-dev/vite): extract route module exports in parallel