--- a/.changeset/fresh-fans-jump.md
+++ b/.changeset/fresh-fans-jump.md
@@ -1,5 +1,0 @@
----
-"@remix-run/dev": patch
----
-
-Run `getRouteModuleExports` in parallel for route manifest generation to improve server startup time