--- a/packages/remix-dev/vite/plugin.ts
+++ b/packages/remix-dev/vite/plugin.ts
@@ -200,6 +200,23 @@ const writeFileSafe = async (file: string, contents: string): Promise<void> => {
   await fse.writeFile(file, contents);
 };
 
+const getRouteManifestModuleExports = async (
+  viteChildCompiler: Vite.ViteDevServer | null,
+  pluginConfig: ResolvedRemixVitePluginConfig
+): Promise<Record<string, string[]>> => {
+  let entries = await Promise.all(
+    Object.entries(pluginConfig.routes).map(async ([key, route]) => {
+      let sourceExports = await getRouteModuleExports(
+        viteChildCompiler,
+        pluginConfig,
+        route.file
+      );
+      return [key, sourceExports] as const;
+    })
+  );
+  return Object.fromEntries(entries);
+};
+
 const getRouteModuleExports = async (
   viteChildCompiler: Vite.ViteDevServer | null,
   pluginConfig: ResolvedRemixVitePluginConfig,
@@ -387,13 +404,15 @@ export const remixVitePlugin: RemixVitePlugin = (options = {}) => {
     );
 
     let routes: Manifest["routes"] = {};
+
+    let routeManifestExports = await getRouteManifestModuleExports(
+      viteChildCompiler,
+      pluginConfig
+    );
+
     for (let [key, route] of Object.entries(pluginConfig.routes)) {
       let routeFilePath = path.join(pluginConfig.appDirectory, route.file);
-      let sourceExports = await getRouteModuleExports(
-        viteChildCompiler,
-        pluginConfig,
-        route.file
-      );
+      let sourceExports = routeManifestExports[key];
 
       routes[key] = {
         id: route.id,
@@ -431,13 +450,13 @@ export const remixVitePlugin: RemixVitePlugin = (options = {}) => {
     let pluginConfig = await resolvePluginConfig();
     let routes: Manifest["routes"] = {};
 
-    for (let [key, route] of Object.entries(pluginConfig.routes)) {
-      let sourceExports = await getRouteModuleExports(
-        viteChildCompiler,
-        pluginConfig,
-        route.file
-      );
+    let routeManifestExports = await getRouteManifestModuleExports(
+      viteChildCompiler,
+      pluginConfig
+    );
 
+    for (let [key, route] of Object.entries(pluginConfig.routes)) {
+      let sourceExports = routeManifestExports[key];
       routes[key] = {
         id: route.id,
         parentId: route.parentId,